5-to_integer.js

+++ 0x12-javascript-warm_up/5-to_integer.js
@@ -1,6 +1,6 @@
 #!/usr/bin/node
 
-const num = parseFloat(process.argv[2], 10);
+const num = parseInt(process.argv[2], 10);
 
 if (!isNaN(num)) {
   console.log('My number: ', num);
